+++ src/components/controlled/transactions-table/components/transaction-account/transaction-account.tsx
@@ -2,38 +2,40 @@ import { faFileAlt } from '@fortawesome/free-solid-svg-icons/faFileAlt';
 import { faLock } from '@fortawesome/free-solid-svg-icons/faLock';
 import { Component, h, Prop } from '@stencil/core';
 import classNames from 'classnames';
+import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
-import type { ITransactionAccount } from '../../transactions-table.type';
+import type { TransactionAccountType } from '../../transactions-table.type';
 
 @Component({
   tag: 'drt-transaction-account',
-  styleUrl: 'transaction-account.css',
+  styleUrl: 'transaction-account.scss',
 })
 export class TransactionAccount {
-  @Prop() account: ITransactionAccount;
+  @Prop() account: TransactionAccountType;
   @Prop() class?: string;
   @Prop() dataTestId?: string;
   @Prop() scope: 'receiver' | 'sender';
   @Prop() showLockedAccounts: boolean = false;
 
   render() {
+    const explorerLinkDataTestId =
+      this.scope === 'receiver' ? DataTestIdsEnum.receiverLink : DataTestIdsEnum.senderLink;
+
     return (
       <div class={classNames(this.class, 'transaction-account')} data-testid={this.dataTestId}>
-        {this.showLockedAccounts && this.account.isTokenLocked && <drt-fa-icon icon={faLock} description={this.account.name}></drt-fa-icon>}
+        {this.showLockedAccounts && this.account.isTokenLocked && (
+          <drt-fa-icon icon={faLock} description={this.account.name} />
+        )}
 
-        {this.account.isContract && <drt-fa-icon icon={faFileAlt} description="Smart Contract"></drt-fa-icon>}
-
+        {this.account.isContract && <drt-fa-icon icon={faFileAlt} description="Smart Contract" />}
         {this.account.showLink ? (
-          <drt-explorer-link link={this.account.link} data-testid={`${this.scope}Link`}>
-            <drt-transaction-account-name
-              slot="content"
-              name={this.account.name}
-              description={this.account.description}
-              address={this.account.address}
-            ></drt-transaction-account-name>
-          </drt-explorer-link>
+          <drt-explorer-link link={this.account.link} data-testid={explorerLinkDataTestId} />
         ) : (
-          <drt-transaction-account-name name={this.account.name} description={this.account.description} address={this.account.address}></drt-transaction-account-name>
+          <drt-transaction-account-name
+            name={this.account.name}
+            description={this.account.description}
+            address={this.account.address}
+          />
         )}
       </div>
     );