+++ src/components/functional/toasts-list/components/transaction-toast/components/transaction-toast-details/transaction-toast-details.tsx
@@ -5,7 +5,7 @@ import type { ITransactionListItem } from 'components/visual/transaction-list-it
 
 @Component({
   tag: 'drt-transaction-toast-details',
-  styleUrl: 'transaction-toast-details.css',
+  styleUrl: 'transaction-toast-details.scss',
 })
 export class TransactionToastDetails {
   @Prop() processedTransactionsStatus?: string | JSX.Element;
@@ -34,12 +34,17 @@ export class TransactionToastDetails {
 
     const hasMoreTransactionsToShow = this.transactions.length > this.maxShownTransactions;
     const hiddenTransactionsCount = this.transactions.length - this.maxShownTransactions;
-    const visibleTransactions = this.showAllTransactions ? this.transactions : this.transactions.slice(0, this.maxShownTransactions);
+    const visibleTransactions = this.showAllTransactions
+      ? this.transactions
+      : this.transactions.slice(0, this.maxShownTransactions);
 
     return (
       <div class="transaction-details-container">
         <div class="transaction-details-status" onClick={this.toggleExpand.bind(this)}>
-          <drt-fa-icon icon={faChevronDown} class={`transaction-details-status-icon ${this.isExpanded ? 'rotate-up' : ''}`}></drt-fa-icon>
+          <drt-fa-icon
+            icon={faChevronDown}
+            class={`transaction-details-status-icon ${this.isExpanded ? 'rotate-up' : ''}`}
+          ></drt-fa-icon>
           <span class="transaction-details-status-text">{this.processedTransactionsStatus}</span>
         </div>
 
@@ -50,7 +55,14 @@ export class TransactionToastDetails {
           }}
         >
           {visibleTransactions.map(({ hash, status, link }, index) => (
-            <drt-transaction-toast-details-body transactionClass={this.transactionClass} hash={hash} status={status} link={link} index={`#${index + 1}`} key={hash} />
+            <drt-transaction-toast-details-body
+              transactionClass={this.transactionClass}
+              hash={hash}
+              status={status}
+              link={link}
+              index={`#${index + 1}`}
+              key={hash}
+            />
           ))}
 
           {hasMoreTransactionsToShow && !this.showAllTransactions && (