+++ src/components/functional/toasts-list/components/transaction-toast/transaction-toast.tsx
@@ -6,7 +6,7 @@ import type { IToastDataState, ITransactionProgressState } from './transaction-t
 
 @Component({
   tag: 'drt-transaction-toast',
-  styleUrl: './transaction-toast.css',
+  styleUrl: './transaction-toast.scss',
 })
 export class TransactionToast {
   @Prop() toastId: string = '';
@@ -25,7 +25,11 @@ export class TransactionToast {
   render() {
     return (
       <div class="transaction-toast">
-        <drt-transaction-toast-progress key={this.toastId} startTime={this.transactionProgressState?.startTime} endTime={this.transactionProgressState?.endTime}>
+        <drt-transaction-toast-progress
+          key={this.toastId}
+          startTime={this.transactionProgressState?.startTime}
+          endTime={this.transactionProgressState?.endTime}
+        >
           <drt-transaction-toast-content
             fullWidth={this.fullWidth}
             toastDataState={this.toastDataState}