+++ src/components/functional/unlock-panel/components/unlock-panel-footer/unlock-panel-footer.tsx
@@ -1,9 +1,10 @@
-import { Component, h } from '@stencil/core';
+import { Component, h, Prop } from '@stencil/core';
 import classNames from 'classnames';
-import { processImgSrc } from 'utils/processImgSrc';
+
+import unlockPanelWalletImg from '../../../../../assets/unlock-panel-wallet.webp';
 
 const unlockPanelClasses: Record<string, string> = {
-  footerIcon: 'drt:w-4! drt:h-auto!',
+  footerIcon: 'drt:w-4! drt:h-auto! drt:hidden drt:xs:flex drt:ml-auto drt:mt-auto',
 };
 
 @Component({
@@ -12,16 +13,34 @@ const unlockPanelClasses: Record<string, string> = {
   shadow: true,
 })
 export class UnlockPanel {
+  @Prop() walletAddress: string;
+
+  handleWalletClick = (event: MouseEvent) => {
+    event.preventDefault();
+    event.stopPropagation();
+    window.open(this.walletAddress, '_blank');
+  };
+
   render() {
+    const processedWalletAddress = String(this.walletAddress).replace('https://', '');
+
     return (
-      <div class="unlock-panel-footer">
-        <img src={processImgSrc('unlock-panel-wallet.png')} class="unlock-panel-footer-image" />
+      <div class="unlock-panel-footer" onClick={this.handleWalletClick}>
+        <img src={unlockPanelWalletImg} class="unlock-panel-footer-image" />
 
         <div class="unlock-panel-footer-wrapper">
           <div class="unlock-panel-footer-title">Don't have a wallet?</div>
 
-          <div class="unlock-panel-footer-subtitle">
+          <div class="unlock-panel-footer-subtitle desktop">
             Take full control of <br /> your assets.
+          </div>
+
+          <div class="unlock-panel-footer-subtitle mobile">
+            <span>See which one to get on </span>
+
+            <a target="_blank" rel="noreferrer" class="unlock-panel-footer-subtitle-link" href={this.walletAddress}>
+              {processedWalletAddress}
+            </a>
           </div>
 
           <drt-arrow-up-right-icon