+++ src/components/visual/explorer-link/tests/explorer-link.spec.ts
@@ -1,4 +1,3 @@
-import { faCheck } from '@fortawesome/free-solid-svg-icons';
 import { newSpecPage } from '@stencil/core/testing';
 
 import { ExplorerLink } from '../explorer-link';
@@ -9,11 +8,14 @@ describe('ExplorerLink', () => {
       components: [ExplorerLink],
       html: '<drt-explorer-link link="https://example.com"></drt-explorer-link>',
     });
-    expect(page.root).toEqualHtml(`
+
+    expect(page.root.outerHTML).toEqualHtml(`
       <drt-explorer-link link="https://example.com">
-        <a href="https://example.com" target="_blank" class="drt:explorer-link" rel="noreferrer">
-          <drt-fa-icon class="explorer-link-icon" icon="faArrowUpRightFromSquare"></drt-fa-icon>
-        </a>
+        <template shadowrootmode="open">
+          <a href="https://example.com" target="_blank" class="explorer-link" rel="noreferrer">
+            <drt-arrow-up-right-from-square-icon class="explorer-link-icon drt:duration-200 drt:ease-in-out drt:flex drt:hover:opacity-80 drt:justify-center drt:transition-opacity"></drt-arrow-up-right-from-square-icon>
+          </a>
+        </template>
       </drt-explorer-link>
     `);
   });
@@ -21,13 +23,17 @@ describe('ExplorerLink', () => {
   it('renders with custom text', async () => {
     const page = await newSpecPage({
       components: [ExplorerLink],
-      html: '<drt-explorer-link link="https://example.com" text="View on Explorer"></drt-explorer-link>',
+      html: '<drt-explorer-link link="https://example.com">View on Explorer</drt-explorer-link>',
     });
-    expect(page.root).toEqualHtml(`
-      <drt-explorer-link link="https://example.com" text="View on Explorer">
-        <a href="https://example.com" target="_blank" class="drt:explorer-link" rel="noreferrer">
-          View on Explorer
-        </a>
+
+    expect(page.root.outerHTML).toEqualHtml(`
+      <drt-explorer-link link="https://example.com">
+        <template shadowrootmode="open">
+          <a href="https://example.com" target="_blank" class="explorer-link" rel="noreferrer">
+            <slot></slot>
+          </a>
+        </template>
+        View on Explorer
       </drt-explorer-link>
     `);
   });
@@ -37,11 +43,14 @@ describe('ExplorerLink', () => {
       components: [ExplorerLink],
       html: '<drt-explorer-link link="https://example.com" class="custom-class"></drt-explorer-link>',
     });
-    expect(page.root).toEqualHtml(`
+
+    expect(page.root.outerHTML).toEqualHtml(`
       <drt-explorer-link link="https://example.com" class="custom-class">
-        <a href="https://example.com" target="_blank" class="drt:explorer-link custom-class" rel="noreferrer">
-          <drt-fa-icon class="explorer-link-icon" icon="faArrowUpRightFromSquare"></drt-fa-icon>
-        </a>
+        <template shadowrootmode="open">
+          <a href="https://example.com" target="_blank" class="explorer-link custom-class" rel="noreferrer">
+            <drt-arrow-up-right-from-square-icon class="explorer-link-icon drt:duration-200 drt:ease-in-out drt:flex drt:hover:opacity-80 drt:justify-center drt:transition-opacity"></drt-arrow-up-right-from-square-icon>
+          </a>
+        </template>
       </drt-explorer-link>
     `);
   });
@@ -49,18 +58,19 @@ describe('ExplorerLink', () => {
   it('renders with custom icon', async () => {
     const page = await newSpecPage({
       components: [ExplorerLink],
-      html: '<drt-explorer-link link="https://example.com"></drt-explorer-link>',
+      html: '<drt-explorer-link link="https://example.com"><drt-arrow-right-icon /></drt-explorer-link>',
     });
 
-    page.root.icon = faCheck;
-    await page.waitForChanges();
-
-    const iconEl = page.root.querySelector('drt-fa-icon');
-    expect(iconEl).toBeTruthy();
-    expect(iconEl.getAttribute('class')).toBe('explorer-link-icon');
-    const iconAttr = iconEl.getAttribute('icon');
-    expect(iconAttr).toContain('<svg');
-    expect(iconAttr).toContain('fa-check');
+    expect(page.root.outerHTML).toEqualHtml(`
+      <drt-explorer-link link="https://example.com">
+        <template shadowrootmode="open">
+          <a href="https://example.com" target="_blank" class="explorer-link" rel="noreferrer">
+            <slot></slot>
+          </a>
+        </template>
+        <drt-arrow-right-icon></drt-arrow-right-icon>
+      </drt-explorer-link>
+    `);
 
     jest.restoreAllMocks();
   });